perf: reduce route calculation time
When the route has 20 or more destinies, use an aproximation to calculate the distance matrix, instead of pathfinding.
Added 3-opt optimization to TSPSolverNN, at the cost of a little more computational time.
But in general, the biggest performance overhead was in the distance matrix calculation using the pathfinder, which is now solved.